brain-workflow-core/pyproject.toml
[project]
name = "brain-workflow-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
